feat(vehicle): 좌표-주소 변환 캐싱 로직 구현 (#68)
- Map 객체를 사용하여 좌표별 주소 데이터 캐싱
- 동일한 좌표 요청 시 캐시된 주소 반환
- 불필요한 API 호출 감소로 성능 개선
------------------------------------------------------

------------------------------------------------------
