+++ app/(dashboard)/location/components/VehicleDetailsCard/index.tsx
@@ -1,12 +1,11 @@
 'use client'
 
 import Image from 'next/image'
-import { useEffect, useState } from 'react'
 
 import Badge from '@/components/common/Badge'
 import SquareButton from '@/components/common/Button/SquareButton'
+import { useCoordToAddress } from '@/hooks/useCoordToAddress'
 import { formatISODateToDot } from '@/lib/utils/date'
-import { convertCoordsToAddress } from '@/lib/utils/map'
 import { vehicleDetailsModel } from '@/types/vehicle'
 
 import * as styles from './styles.css'
@@ -17,8 +16,6 @@ interface VehicleDetailsCardProps {
 }
 
 const VehicleDetailsCard = ({ vehicleDetails, onCloseButtonClick }: VehicleDetailsCardProps) => {
-    const [address, setAddress] = useState('')
-
     const {
         department,
         vehicleNumber,
@@ -28,13 +25,7 @@ const VehicleDetailsCard = ({ vehicleDetails, onCloseButtonClick }: VehicleDetai
         location: { lat, lng, lastUpdated },
     } = vehicleDetails
 
-    useEffect(() => {
-        const getAddressFromCoords = async () => {
-            const address = await convertCoordsToAddress(lat, lng)
-            setAddress(address)
-        }
-        getAddressFromCoords()
-    }, [lat, lng])
+    const address = useCoordToAddress(lat, lng)
 
     const isDriving = type === 'ON'
 
@@ -106,7 +97,6 @@ const VehicleDetailsCard = ({ vehicleDetails, onCloseButtonClick }: VehicleDetai
                                 최종위치
                             </th>
                             <td colSpan={3} className={styles.tableCell}>
-                                {/* 대구광역시 동구 효목동 209 */}
                                 {address}
                             </td>
                         </tr>
